requirements.txt
# Core dependencies
ibm-db>=3.2.0
python-dotenv>=1.0.0
requests>=2.31.0

# Document processing
docling>=2.0.0
langchain-text-splitters>=0.3.0

# Web dashboard
dash>=2.18.0
dash-table>=5.0.0

# Optional: For development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0